// apps/mobile/src/api/generated/kitchenUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen API
 * OpenAPI spec version: 0.1.0
 */

export interface KitchenUpdate {
  zone_id?: string | null;
  address_label?: string | null;
  name?: string | null;
  description?: string | null;
  pickup_enabled?: boolean | null;
  delivery_enabled?: boolean | null;
  delivery_fee_paise?: number | null;
  upi_id?: string | null;
  fssai_number?: string | null;
}