// apps/mobile/src/api/generated/listingCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingCreateStatus } from './listingCreateStatus';

export interface ListingCreate {
  /** @maxItems 30 */
  pickup_point_ids?: string[];
  dish_id: string;
  service_date: string;
  available_from: string;
  available_until: string;
  order_cutoff: string;
  /**
     * @maximum 10000000
     * @exclusiveMinimum 0
     */
  price_paise: number;
  /**
     * @maximum 10000
     * @exclusiveMinimum 0
     */
  quantity_total: number;
  pickup_enabled?: boolean;
  delivery_enabled?: boolean;
  status?: ListingCreateStatus;
}